snapshot_convertor: move the gzip utils to their own trait
